Re-arrange tools and include in Microsoft.Health.Dicom.sln
