+++ tools/dicom-auth-test/Image/blue-circle.dcm
