+++ tools/dicom-auth-test/Microsoft.Health.Dicom.Tool.ManagedIdentitySample.csproj
@@ -6,20 +6,18 @@
     <Nullable>enable</Nullable>
     <OutputType>Exe</OutputType>
     <NoWarn>$(NoWarn);NU5104</NoWarn>
+    <RootNamespace>Microsoft.Health.Web.Dicom.Tool</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-
-    <AssemblyAttribute Include="System.CLSCompliant">
-      <_Parameter1>true</_Parameter1>
-      <_Parameter1_IsLiteral>true</_Parameter1_IsLiteral>
-    </AssemblyAttribute>
     <PackageReference Include="Azure.Core" />
     <PackageReference Include="Azure.Identity" />
     <PackageReference Include="fo-dicom" />
-    <PackageReference Include="Microsoft.Health.Dicom.Client" />
-    <PackageReference Include="System.CommandLine"/>
     <PackageReference Include="Newtonsoft.Json" />
+    <PackageReference Include="System.CommandLine" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Microsoft.Health.Dicom.Client\Microsoft.Health.Dicom.Client.csproj" />
+  </ItemGroup>
 </Project>
