--- tools/dicom-auth-test/Microsoft.Health.Web.Dicom.Tool/Microsoft.Health.Web.Dicom.Tool.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.2.32505.173
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Health.Web.Dicom.Tool", "Microsoft.Health.Web.Dicom.Tool.csproj", "{FD5481F9-E9AE-498B-A3AD-901D17B44E02}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FD5481F9-E9AE-498B-A3AD-901D17B44E02}.Debug|x64.ActiveCfg = Debug|x64
-		{FD5481F9-E9AE-498B-A3AD-901D17B44E02}.Debug|x64.Build.0 = Debug|x64
-		{FD5481F9-E9AE-498B-A3AD-901D17B44E02}.Release|x64.ActiveCfg = Release|x64
-		{FD5481F9-E9AE-498B-A3AD-901D17B44E02}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C565C52C-7840-4403-B06F-D85247B28143}
-	EndGlobalSection
-EndGlobal
