+++ tools/dicom-auth-test/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿// -------------------------------------------------------------------------------------------------
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+// -------------------------------------------------------------------------------------------------
+
+[assembly: CLSCompliant(false)]
