+++ tools/uploader-function/Configuration/DicomOptions.cs
