+++ tools/uploader-function/Dockerfile
