+++ tools/uploader-function/Properties/AssemblyInfo.cs
