+++ tools/uploader-function/Startup.cs
